dice-protocols: maudio: move structure for hardware specification
This commit moves implementation of hardware specification for M-Audio
ProFire series from runtime crate to protocol crate.
