+++ libs/dice/runtime/src/pfire_model.rs
@@ -144,104 +144,6 @@ impl<S> MeasureModel<hinawa::SndDice> for PfireModel<S>
     }
 }
 
-pub trait PfireClkSpec<'a> {
-    const AVAIL_CLK_RATES: &'a [ClockRate] = &[
-        ClockRate::R32000, ClockRate::R44100, ClockRate::R48000,
-        ClockRate::R88200, ClockRate::R96000,
-        ClockRate::R176400, ClockRate::R192000,
-    ];
-
-    const AVAIL_CLK_SRCS: &'a [ClockSource];
-}
-
-#[derive(Default, Debug)]
-pub struct Pfire2626State(Tcd22xxState);
-
-impl<'a> Tcd22xxSpec<'a> for Pfire2626State {
-    const INPUTS: &'a [Input<'a>] = &[
-        Input{id: SrcBlkId::Ins1, offset: 0, count: 8, label: None},
-        Input{id: SrcBlkId::Adat, offset: 0, count: 8, label: None},
-        Input{id: SrcBlkId::Adat, offset: 8, count: 16, label: None},
-        Input{id: SrcBlkId::Aes, offset: 0, count: 2, label: None},
-    ];
-    const OUTPUTS: &'a [Output<'a>] = &[
-        Output{id: DstBlkId::Ins1, offset: 0, count: 8, label: None},
-        Output{id: DstBlkId::Adat, offset: 0, count: 8, label: None},
-        Output{id: DstBlkId::Adat, offset: 8, count: 16, label: None},
-        Output{id: DstBlkId::Aes, offset: 0, count: 2, label: None},
-    ];
-    const FIXED: &'a [SrcBlk] = &[
-        SrcBlk{id: SrcBlkId::Ins1, ch: 0},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 1},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 2},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 3},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 4},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 5},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 6},
-        SrcBlk{id: SrcBlkId::Ins1, ch: 7},
-    ];
-}
-
-impl AsMut<Tcd22xxState> for Pfire2626State {
-    fn as_mut(&mut self) -> &mut Tcd22xxState {
-        &mut self.0
-    }
-}
-
-impl AsRef<Tcd22xxState> for Pfire2626State {
-    fn as_ref(&self) -> &Tcd22xxState {
-        &self.0
-    }
-}
-
-impl<'a> PfireClkSpec<'a> for Pfire2626State {
-    const AVAIL_CLK_SRCS: &'a [ClockSource] = &[
-            ClockSource::Aes1,
-            ClockSource::Aes4,
-            ClockSource::Adat,
-            ClockSource::Tdif,
-            ClockSource::WordClock,
-            ClockSource::Internal,
-    ];
-}
-
-#[derive(Default, Debug)]
-pub struct Pfire610State(Tcd22xxState);
-
-impl<'a> Tcd22xxSpec<'a> for Pfire610State {
-    const INPUTS: &'a [Input<'a>] = &[
-        Input{id: SrcBlkId::Ins0, offset: 0, count: 4, label: None},
-        Input{id: SrcBlkId::Aes,  offset: 0, count: 2, label: None},
-    ];
-    const OUTPUTS: &'a [Output<'a>] = &[
-        Output{id: DstBlkId::Ins0, offset: 0, count: 8, label: None},
-        Output{id: DstBlkId::Aes,  offset: 0, count: 2, label: None},
-    ];
-    const FIXED: &'a [SrcBlk] = &[
-        SrcBlk{id: SrcBlkId::Ins0, ch: 0},
-        SrcBlk{id: SrcBlkId::Ins0, ch: 1},
-    ];
-}
-
-impl AsRef<Tcd22xxState> for Pfire610State {
-    fn as_ref(&self) -> &Tcd22xxState {
-        &self.0
-    }
-}
-
-impl AsMut<Tcd22xxState> for Pfire610State {
-    fn as_mut(&mut self) -> &mut Tcd22xxState {
-        &mut self.0
-    }
-}
-
-impl<'a> PfireClkSpec<'a> for Pfire610State {
-    const AVAIL_CLK_SRCS: &'a [ClockSource] = &[
-            ClockSource::Aes1,
-            ClockSource::Internal,
-    ];
-}
-
 #[derive(Default, Debug)]
 struct SpecificCtl{
     targets: [bool;KNOB_COUNT],
